+++ client/.env
@@ -1,0 +1,2 @@
+REACT_APP_MAPS_API_KEY=AIzaSyDppjGMn8dVBBREFUOAYIA15foyW3dUgT4
+REACT_APP_MAP_ID=2a04f071da976688
