+++ client/package.json
@@ -15,6 +15,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.12.0",
+    "react-router-dom": "^6.20.0",
     "react-scripts": "5.0.1",
     "use-places-autocomplete": "^4.0.1",
     "web-vitals": "^2.1.4"
