+++ client/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 // import ReactDOM from 'react-dom/client';
-import App from './App';
 import { createRoot } from 'react-dom/client'
+import App from './App';
 
 
 const root = createRoot(document.getElementById('root'));
